don't automatically set error state from unconfirmed
when using a service like infura to send transactions, the block
monitor's node wont always see the unconfirmed transactions before
they're mined, and if these took longer than the 2 minutes to mine
this would set them as an error, while they would eventually be mined
anyways.

planning to replace this with something to better report errors for
now and if they happen try and come up with a better way to deal with them.